moved the TriangleApplication class to its own source file
