+++ vulkan_window/triangle_application.cpp
@@ -1,0 +1,407 @@
+/* Local header files */
+#include "triangle_application.hpp"
+
+bool TriangleApplication::QueueFamilyIndices::isComplete()
+{
+    // Checks if the graphicsFamily object contains a value
+    return graphicsFamily.has_value();
+}
+
+void TriangleApplication::run()
+{
+    initWindow();
+    initVulkan();
+    mainLoop();
+    cleanUp();
+}
+
+void TriangleApplication::initWindow()
+{
+    /* Initialize the GLFW window */
+    // Initialize GLFW libary
+    glfwInit();
+
+    // Inform GLFW to not create an OpenGL context
+    // GLFW was originally designed to create an OpenGL context
+    glfwWindowHint(GLFW_CLIENT_API, GLFW_NO_API);
+
+    // Disable handling resized windows
+    glfwWindowHint(GLFW_RESIZABLE, GLFW_FALSE);
+
+    // Create the GLFW window
+    window = glfwCreateWindow(WIDTH, HEIGHT, "Vulkan window", nullptr, nullptr);
+}
+
+void TriangleApplication::initVulkan()
+{
+    /* Initialize Vulkan */
+    createInstance();
+    setupDebugMessenger();
+    pickPhysicalDevice();
+    createLogicalDevice();
+}
+
+void TriangleApplication::mainLoop()
+{
+    /* Main game loop */
+    while (!glfwWindowShouldClose(window))
+    {
+        glfwPollEvents();
+    }
+}
+
+void TriangleApplication::cleanUp()
+{
+    /* Clean up resources */
+    vkDestroyDevice(device, nullptr);
+
+    if (enableValidationLayers)
+    {
+        DestroyDebugUtilsMessengerEXT(instance, debugMessenger, nullptr);
+    }
+
+    vkDestroyInstance(instance, nullptr);
+
+    glfwDestroyWindow(window);
+
+    glfwTerminate();
+}
+
+void TriangleApplication::checkExtensionSupport()
+{
+    /* Checking for extension support */
+    // Count the amount of supported extensions
+    uint32_t extensionCount = 0;
+    vkEnumerateInstanceExtensionProperties(nullptr, &extensionCount, nullptr);
+
+    // Query extension details
+    std::vector<VkExtensionProperties> extensions(extensionCount);
+    vkEnumerateInstanceExtensionProperties(nullptr, &extensionCount, extensions.data());
+
+    std::cout << "available extensions: " << std::endl;
+    for (const auto& extension : extensions)
+    {
+        std::cout << '\t' << extension.extensionName << '\n';
+    }
+}
+
+void TriangleApplication::createInstance()
+{
+    /* Create instance */
+    if (enableValidationLayers && !checkValidationLayerSupport())
+    {
+        throw std::runtime_error("validation layers requested, but not available!");
+    }
+
+    // Fill in some information about the application
+    // This data is optional.
+    VkApplicationInfo appInfo{};
+    appInfo.sType = VK_STRUCTURE_TYPE_APPLICATION_INFO;
+    appInfo.pApplicationName = "Triangle";
+    appInfo.applicationVersion = VK_MAKE_VERSION(0, 1, 0);
+    appInfo.pEngineName = "No Engine";
+    appInfo.engineVersion = VK_MAKE_VERSION(0, 1, 0);
+    appInfo.apiVersion = VK_API_VERSION_1_0;
+
+    // Informs the Vulkan driver which global extensions and
+    // validation layers we want to use
+    VkInstanceCreateInfo createInfo{};
+    createInfo.sType = VK_STRUCTURE_TYPE_INSTANCE_CREATE_INFO;
+    createInfo.pApplicationInfo = &appInfo;
+
+    // Add an extension to inferface with the window system for GLFW
+    uint32_t glfwExtensionCount = 0;
+    const char** glfwExtensions;
+
+    // This function returns an array of required Vulkan instance extensions
+    // for creating Vulkan surfaces on GLFW windows
+    glfwExtensions = glfwGetRequiredInstanceExtensions(&glfwExtensionCount);
+
+    createInfo.enabledExtensionCount = glfwExtensionCount;
+    createInfo.ppEnabledExtensionNames = glfwExtensions;
+
+    // Retreive the required list of extensions
+    auto extensions = getRequiredExtensions();
+    createInfo.enabledExtensionCount = static_cast<uint32_t>(extensions.size());
+    createInfo.ppEnabledExtensionNames = extensions.data();
+
+    VkDebugUtilsMessengerCreateInfoEXT debugCreateInfo{};
+
+    // Modify the VkInstanceCreateInfo struct to include the validation layer names
+    // if they are enabled
+    if (enableValidationLayers)
+    {
+        createInfo.enabledLayerCount = static_cast<uint32_t>(validationLayers.size());
+        createInfo.ppEnabledLayerNames = validationLayers.data();
+
+        populateDebugMessengerCreateInfo(debugCreateInfo);
+        createInfo.pNext = (VkDebugUtilsMessengerCreateInfoEXT*)&debugCreateInfo;
+    }
+    else
+    {
+        createInfo.enabledLayerCount = 0;
+        createInfo.pNext = nullptr;
+    }
+
+    // Create an instance
+    if (vkCreateInstance(&createInfo, nullptr, &instance) != VK_SUCCESS)
+    {
+        throw std::runtime_error("vkCreateInstance ERROR: failed to create instance!");
+    }
+
+    // checkExtensionSupport();
+}
+
+bool TriangleApplication::checkValidationLayerSupport()
+{
+    /* Checks if all of the requested layers are available */
+    // List all of the available layers
+    uint32_t layerCount;
+    vkEnumerateInstanceLayerProperties(&layerCount, nullptr);
+
+    std::vector<VkLayerProperties> availableLayers(layerCount);
+    vkEnumerateInstanceLayerProperties(&layerCount, availableLayers.data());
+
+    // Check if all of the layers in the validationLayers exist in the
+    // availableLayers list
+    for (const char* layerName : validationLayers)
+    {
+        bool layerFound = false;
+
+        for (const auto& layerProperties : availableLayers)
+        {
+            if (strcmp(layerName, layerProperties.layerName) == 0)
+            {
+                layerFound = true;
+                break;
+            }
+        }
+
+        if (!layerFound)
+        {
+            return false;
+        }
+    }
+
+    return true;
+}
+
+std::vector<const char*> TriangleApplication::getRequiredExtensions()
+{
+    /* Retreive the required list of extensions based on if the
+    validation layers are enabled or disabled */
+    uint32_t glfwExtensionCount = 0;
+    const char** glfwExtensions;
+    glfwExtensions = glfwGetRequiredInstanceExtensions(&glfwExtensionCount);
+
+    std::vector<const char*> extensions(glfwExtensions, glfwExtensions + glfwExtensionCount);
+
+    if (enableValidationLayers)
+    {
+        extensions.push_back(VK_EXT_DEBUG_UTILS_EXTENSION_NAME);
+    }
+
+    return extensions;
+}
+
+VKAPI_ATTR VkBool32 VKAPI_CALL TriangleApplication::debugCallback(
+    VkDebugUtilsMessageSeverityFlagBitsEXT messageSeverity,
+    VkDebugUtilsMessageTypeFlagsEXT messageType,
+    const VkDebugUtilsMessengerCallbackDataEXT* pCallbackData,
+    void* pUserData)
+{
+    /* Debug to the console for validation layers */
+    std::string debug_msg = "validation layer: " + (std::string)pCallbackData->pMessage;
+    std::cerr << debug_msg << std::endl;
+    OutputDebugStringA(debug_msg.c_str());
+
+    if (messageSeverity >= VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT)
+    {
+        // Message is important enought to show
+        debug_msg = "validation layer ERROR: " + (std::string)pCallbackData->pMessage;
+        throw std::runtime_error(debug_msg.c_str());
+    }
+
+    return VK_FALSE;
+}
+
+void TriangleApplication::setupDebugMessenger()
+{
+    if (!enableValidationLayers)
+    {
+        return;
+    }
+
+    // Fill in the details about the messenger and its callback
+    VkDebugUtilsMessengerCreateInfoEXT createInfo{};
+    populateDebugMessengerCreateInfo(createInfo);
+
+    // Create the extension object if it is available
+    if (CreateDebugUtilsMessengerEXT(instance, &createInfo, nullptr, &debugMessenger) != VK_SUCCESS)
+    {
+        throw std::runtime_error("failed to set up debug messenger!");
+    }
+}
+
+VkResult TriangleApplication::CreateDebugUtilsMessengerEXT(VkInstance instance,
+    const VkDebugUtilsMessengerCreateInfoEXT* pCreateInfo,
+    const VkAllocationCallbacks* pAllocator, VkDebugUtilsMessengerEXT* pDebugMessenger)
+{
+    /* Proxy function to create the debug messenger */
+    // Looks up the address of the VkDebugUtilsMessengerEXT object
+    auto func = (PFN_vkCreateDebugUtilsMessengerEXT)vkGetInstanceProcAddr(instance, "vkCreateDebugUtilsMessengerEXT");
+
+    if (func != nullptr)
+    {
+        return func(instance, pCreateInfo, pAllocator, pDebugMessenger);
+    }
+    else
+    {
+        return VK_ERROR_EXTENSION_NOT_PRESENT;
+    }
+}
+
+void TriangleApplication::DestroyDebugUtilsMessengerEXT(VkInstance instance,
+    VkDebugUtilsMessengerEXT debugMessenger,
+    const VkAllocationCallbacks* pAllocator)
+{
+    /* Proxy function to destroy the debug messenger */
+    auto func = (PFN_vkDestroyDebugUtilsMessengerEXT)vkGetInstanceProcAddr(instance,
+        "vkDestroyDebugUtilsMessengerEXT");
+
+    if (func != nullptr)
+    {
+        func(instance, debugMessenger, pAllocator);
+    }
+}
+
+void TriangleApplication::populateDebugMessengerCreateInfo(VkDebugUtilsMessengerCreateInfoEXT& createInfo)
+{
+    /* Fill in the details about the messenger and its callback to the structure */
+    createInfo = {};
+    createInfo.sType = VK_STRUCTURE_TYPE_DEBUG_UTILS_MESSENGER_CREATE_INFO_EXT;
+    createInfo.messageSeverity = VK_DEBUG_UTILS_MESSAGE_SEVERITY_VERBOSE_BIT_EXT |
+        VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT |
+        VK_DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT;
+    createInfo.messageType = VK_DEBUG_UTILS_MESSAGE_TYPE_GENERAL_BIT_EXT |
+        VK_DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT |
+        VK_DEBUG_UTILS_MESSAGE_TYPE_PERFORMANCE_BIT_EXT;
+    createInfo.pfnUserCallback = debugCallback;
+}
+
+void TriangleApplication::pickPhysicalDevice()
+{
+    // Count the number of graphics cards with Vulkan support
+    uint32_t deviceCount = 0;
+    vkEnumeratePhysicalDevices(instance, &deviceCount, nullptr);
+
+    if (deviceCount == 0)
+    {
+        throw std::runtime_error("failed to find GPUs with Vulkan support!");
+    }
+
+    // Allocate an array to hold all of the VkPhysicalDevice handles
+    std::vector<VkPhysicalDevice> devices(deviceCount);
+    vkEnumeratePhysicalDevices(instance, &deviceCount, devices.data());
+
+    // Device suitability checks
+    for (const auto& device : devices)
+    {
+        if (isDeviceSuitable(device))
+        {
+            physicalDevice = device;
+            break;
+        }
+    }
+
+    if (physicalDevice == VK_NULL_HANDLE)
+    {
+        throw std::runtime_error("failed to find a suitable GPU!");
+    }
+}
+
+bool TriangleApplication::isDeviceSuitable(VkPhysicalDevice device)
+{
+    /* Queue family lookup function to ensure the device can process the
+    commands to use */
+    QueueFamilyIndices indices = findQueueFamilies(device);
+
+    return indices.isComplete();
+}
+
+TriangleApplication::QueueFamilyIndices TriangleApplication::findQueueFamilies(VkPhysicalDevice device)
+{
+    /* Logic to find queue family indices to populate the struct with */
+    QueueFamilyIndices indices;
+
+    // The process to retreive the list of queue families
+    uint32_t queueFamilyCount = 0;
+    vkGetPhysicalDeviceQueueFamilyProperties(device, &queueFamilyCount, nullptr);
+
+    std::vector<VkQueueFamilyProperties> queueFamilies(queueFamilyCount);
+    vkGetPhysicalDeviceQueueFamilyProperties(device, &queueFamilyCount, queueFamilies.data());
+
+    // Find at lest one queue family that supports VK_QUEUE_GRAPHICS_BIT
+    int i = 0;
+    for (const auto& queueFamily : queueFamilies)
+    {
+        if (queueFamily.queueFlags & VK_QUEUE_GRAPHICS_BIT)
+        {
+            indices.graphicsFamily = i;
+        }
+
+        if (indices.isComplete())
+        {
+            break;
+        }
+
+        i++;
+    }
+
+    return indices;
+}
+
+void TriangleApplication::createLogicalDevice()
+{
+    // Specify the queues to be created
+    QueueFamilyIndices indicies = findQueueFamilies(physicalDevice);
+
+    VkDeviceQueueCreateInfo queueCreateInfo{};
+    queueCreateInfo.sType = VK_STRUCTURE_TYPE_DEVICE_QUEUE_CREATE_INFO;
+    queueCreateInfo.queueFamilyIndex = indicies.graphicsFamily.value();
+    queueCreateInfo.queueCount = 1;
+
+    float queuePriority = 1.0f;
+    queueCreateInfo.pQueuePriorities = &queuePriority;
+
+    // Specify used device features
+    VkPhysicalDeviceFeatures deviceFeatures{};
+
+    // Create the logical device
+    VkDeviceCreateInfo createInfo{};
+    createInfo.sType = VK_STRUCTURE_TYPE_DEVICE_CREATE_INFO;
+    createInfo.pQueueCreateInfos = &queueCreateInfo;
+    createInfo.queueCreateInfoCount = 1;
+    createInfo.pEnabledFeatures = &deviceFeatures;
+
+    // Specify the extensions and validation layers
+    createInfo.enabledExtensionCount = 0;
+    if (enableValidationLayers)
+    {
+        createInfo.enabledLayerCount = static_cast<uint32_t>(validationLayers.size());
+        createInfo.ppEnabledLayerNames = validationLayers.data();
+    }
+    else
+    {
+        createInfo.enabledLayerCount = 0;
+    }
+
+    // Instantiate the logical device
+    if (vkCreateDevice(physicalDevice, &createInfo, nullptr, &device) != VK_SUCCESS)
+    {
+        throw std::runtime_error("failed to create logical device!");
+    }
+
+    // Retreving queue handles
+    vkGetDeviceQueue(device, indicies.graphicsFamily.value(), 0, &graphicsQueue);
+}
