+++ vulkan_window/triangle_application.hpp
@@ -1,0 +1,87 @@
+#ifndef TRIANGLE_APPLICATION_H
+#define TRIANGLE_APPLICATION_H
+
+/* Third party libraries */
+#include <vulkan/vulkan.h>
+
+#define GLFW_INCLUDE_VULKAN
+#include <GLFW/glfw3.h>
+
+#define GLM_FORCE_RADIUS
+#define GLM_FORCE_DEPTH_ZERO_TO_ONE
+#include <vec4.hpp>
+#include <mat4x4.hpp>
+
+/* Standard libraries */
+#include <iostream>
+#include <stdexcept>
+#include <cstdlib>
+#include <vector>
+#include <cstring>
+#include <Windows.h>
+#include <map>
+#include <optional>
+
+const uint32_t WIDTH = 800;
+const uint32_t HEIGHT = 600;
+
+// Enable the standard diagnostic layers provided by the Vulkan SDK
+const std::vector<const char*> validationLayers = {
+    "VK_LAYER_KHRONOS_validation" };
+
+#define NDEBUG
+
+#ifndef NDEBUG // not debug
+const bool enableValidationLayers = false;
+#else
+const bool enableValidationLayers = true;
+#endif
+
+class TriangleApplication
+{
+    GLFWwindow* window;
+    VkInstance instance;
+    VkDebugUtilsMessengerEXT debugMessenger;
+    VkPhysicalDevice physicalDevice = VK_NULL_HANDLE;
+    VkDevice device;
+    VkQueue graphicsQueue;
+
+    struct QueueFamilyIndices
+    {
+        std::optional<uint32_t> graphicsFamily;
+
+        bool isComplete();
+    };
+
+public:
+    void run();
+
+private:
+    void initWindow();
+    void initVulkan();
+    void mainLoop();
+    void cleanUp();
+    void checkExtensionSupport();
+    void createInstance();
+    bool checkValidationLayerSupport();
+    std::vector<const char*> getRequiredExtensions();
+    static VKAPI_ATTR VkBool32 VKAPI_CALL debugCallback(
+        VkDebugUtilsMessageSeverityFlagBitsEXT messageSeverity,
+        VkDebugUtilsMessageTypeFlagsEXT messageType,
+        const VkDebugUtilsMessengerCallbackDataEXT* pCallbackData,
+        void* pUserData);
+    void setupDebugMessenger();
+    VkResult CreateDebugUtilsMessengerEXT(VkInstance instance,
+        const VkDebugUtilsMessengerCreateInfoEXT* pCreateInfo,
+        const VkAllocationCallbacks* pAllocator, VkDebugUtilsMessengerEXT* pDebugMessenger);
+    void DestroyDebugUtilsMessengerEXT(VkInstance instance,
+        VkDebugUtilsMessengerEXT debugMessenger,
+        const VkAllocationCallbacks* pAllocator);
+    void populateDebugMessengerCreateInfo(VkDebugUtilsMessengerCreateInfoEXT& createInfo);
+    void pickPhysicalDevice();
+    bool isDeviceSuitable(VkPhysicalDevice device);
+    QueueFamilyIndices findQueueFamilies(VkPhysicalDevice device);
+    void createLogicalDevice();
+};
+
+#endif // TRIANGLE_APPLICATION_H
